Move collectInstantiatedModules into private util function
